+++ Plugins/ActorPools/Source/NexusActorPools/NexusActorPools.Build.cs
@@ -20,6 +20,7 @@ namespace UnrealBuildTool.Rules
 			PrivateDependencyModuleNames.AddRange(
 				[
 					"CoreUObject",
+					"DeveloperSettings",
 					"Projects"
 				]
 			);